@@ -26,41 +26,41 @@
     "examples": "./examples/build.sh",
     "lint": "eslint lib test",
     "test": "npm run lint && jasmine JASMINE_CONFIG_PATH=test/jasmine.json",
-    "prepublish": "npm run test"
+    "prepublishOnly": "npm run test"
   },
   "dependencies": {
     "hash-sum": "^1.0.2",
     "loader-utils": "^1.1.0",
-    "postcss": "^5.2.16",
-    "postcss-selector-parser": "^2.2.3",
-    "source-map": "^0.5.6",
-    "vue-hot-reload-api": "^2.0.11",
-    "vue-template-es2015-compiler": "^1.5.1"
+    "postcss": "^6.0.16",
+    "postcss-selector-parser": "^3.1.1",
+    "source-map": "^0.6.1",
+    "vue-hot-reload-api": "^2.2.4",
+    "vue-template-es2015-compiler": "^1.6.0"
   },
   "devDependencies": {
-    "babel-eslint": "^7.2.1",
-    "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-eslint": "^8.2.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-register": "^6.24.0",
-    "css-loader": "^0.28.0",
-    "eslint": "^3.19.0",
+    "babel-register": "^6.26.0",
+    "css-loader": "^0.28.8",
+    "eslint": "^4.15.0",
     "eslint-config-ktsn": "^1.0.3",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.11.1",
-    "html-webpack-plugin": "^2.28.0",
-    "jasmine": "^2.5.3",
-    "js-base64": "^2.1.9",
-    "node-sass": "^4.5.2",
-    "pug-html-loader": "^1.1.4",
-    "require-from-string": "^1.2.1",
-    "sass-loader": "^6.0.3",
-    "style-loader": "^0.16.1",
-    "ts-loader": "^2.0.3",
-    "typescript": "^2.2.2",
-    "vue": "^2.2.6",
-    "vue-class-component": "^5.0.1",
-    "vue-template-compiler": "^2.2.6",
-    "webpack": "^2.3.2"
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.6",
+    "html-webpack-plugin": "^2.30.1",
+    "jasmine": "^2.8.0",
+    "js-base64": "^2.4.0",
+    "node-sass": "^4.7.2",
+    "pug-html-loader": "^1.1.5",
+    "require-from-string": "^2.0.1",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.1",
+    "ts-loader": "^3.2.0",
+    "typescript": "^2.6.2",
+    "vue": "^2.5.13",
+    "vue-class-component": "^6.1.2",
+    "vue-template-compiler": "^2.5.13",
+    "webpack": "^3.10.0"
   },
   "peerDependencies": {
     "vue-template-compiler": "^2.0.0"
