@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import Component from 'vue-class-component'
-import * as Template from './app.pug?style=./app.scss'
+import Template from './app.pug?style=./app.scss'
 
 @Template
 @Component
