@@ -68,7 +68,7 @@ const addScopedIdPlugin = postcss.plugin('add-scoped-id', options => {
           node.each(transformNode)
         }
       } else {
-        node.selector = selectorTransformer.process(node.selector).result
+        node.selector = selectorTransformer.processSync(node.selector)
       }
     })
 
